Moves event api into an event controller.
